02. Счастливые билеты (массивы)

+++ src/main/java/ru/algo/tickets/HappyTicketsRecursive.java
@@ -13,21 +13,21 @@ public class HappyTicketsRecursive implements CommonAlgo {
 
     @Override
     public String[] exec(String[] input) {
-        int n = parseInt(input[0]);
+        long n = parseInt(input[0]);
 
-        return new String[] { Integer.toString(findTickets(n)) };
+        return new String[] { Long.toString(findTickets(n)) };
     }
 
     // ======= Алгоритм =======
-    private int findTickets(int n) {
+    private long findTickets(long n) {
         algo(n, 0, 0);
 
         return count;
     }
 
-    private int count = 0;
+    private long count = 0;
 
-    private void algo(int remainingDigits, int sumA, int sumB) {
+    private void algo(long remainingDigits, long sumA, long sumB) {
 
         if (remainingDigits == 0) {
             if (sumA == sumB)
@@ -36,8 +36,8 @@ public class HappyTicketsRecursive implements CommonAlgo {
             return;
         }
 
-        for (int a = 0; a <= 9; a++)
-            for (int b = 0; b <= 9; b++)
+        for (long a = 0; a <= 9; a++)
+            for (long b = 0; b <= 9; b++)
                 algo(remainingDigits - 1, sumA + a, sumB + b);
     }
 }
